+++ src/modules/main/mnemonic/MnemonicComponent.tsx
@@ -1,7 +1,7 @@
 import React, {FC, useState} from "react";
 import s from "../EngWord.module.css";
 import s2 from "./mnemonic.module.css";
-import {IMnemonic} from "../../../shared/models/engWordTypes";
+import {IEngWord, IMnemonic} from "../../../shared/models/engWordTypes";
 import HighlightMnemonic from "./HighlightMnemonic";
 import {useAppDispatch} from "../../../store";
 import {
@@ -20,10 +20,15 @@ import {ReactComponent as Minus} from "../../../import/icons/minus.svg"
 import {ReactComponent as Bold} from "../../../import/icons/bold1.svg"
 import {ReactComponent as NotBold} from "../../../import/icons/bold2.svg"
 import UpdateMnemonic from "./UpdateMnemonic";
-
+import {DateAgo} from "../DateAgo";
+import {useNavigate} from "react-router";
+import {DeleteModal} from "../Modal/DeleteModal";
 
 type PropsType = {
     mnemonic: IMnemonic
+    auth: boolean
+    engWord: IEngWord
+
 }
 const MnemonicComponent: FC<PropsType> = (props) => {
     const [bold, setBold] = useState<boolean>(false);
@@ -31,16 +36,31 @@ const MnemonicComponent: FC<PropsType> = (props) => {
     let pushBold = () => {
         setBold(!bold)
     };
+    const [showModal, setShowModal] = useState(false)
+    const closeDeleteModal = () => {
+        setShowModal(false)
+    }
     const dispatch = useAppDispatch();
+   const navigate = useNavigate()
     let addLikeMnemonic = () => {
-        if (!props.mnemonic.isLiked) {
-            dispatch(addMnemonicLikeAsync(props.mnemonic.mnemonicId))
+        if (props.auth) {
+            if (!props.mnemonic.isLiked) {
+                dispatch(addMnemonicLikeAsync(props.mnemonic.mnemonicId))
+            } else {
+                dispatch(deleteMnemonicLikeAsync(props.mnemonic.mnemonicId))
+            }
         } else {
-            dispatch(deleteMnemonicLikeAsync(props.mnemonic.mnemonicId))
+            navigate(`/login`)
         }
+
     };
+    const showDeleteModal = () => {
+        setShowModal(true)
+    }
+
     let deleteMnemonic = () => {
         dispatch(deleteMnemonicAsync(props.mnemonic.mnemonicId))
+        setShowModal(false)
     };
     let addMeMnemonic = () => {
         if (!props.mnemonic.added) {
@@ -62,20 +82,36 @@ const MnemonicComponent: FC<PropsType> = (props) => {
     if (props.mnemonic.added) {
         addDelete = <Minus/>;
     }
-
     let boldIcon = <NotBold/>;
         if(bold){
             boldIcon = <Bold/>
+        }
+        let date = (date: string) => {
+            return (
+                <div>
+                    <DateAgo date={date}/>
+                </div>
+            )
         }
 
     return (
 
         <div className={s.mnemonicComponent}>
+
             {!edit ?
                 <>
-                <div className={s.highlightWord}>
-                    <b>{props.mnemonic.highlightWord}</b>
-                </div>
+                    <div className={s.mnemonicBottom}>
+                        <div className={s.highlightWord}>
+                            <b>{props.mnemonic.highlightWord}</b>
+                        </div>
+                        <div className={s.accuracy}>
+                            {Math.round(props.mnemonic.accuracy)}%
+                        </div>
+                        <div className={s.dateFormat}>
+                            {date(props.mnemonic.created)}
+                        </div>
+                    </div>
+
 
                 <div className={s.mnemo}>
                     {bold ? <HighlightMnemonic
@@ -88,15 +124,22 @@ const MnemonicComponent: FC<PropsType> = (props) => {
                     <span className={s.author}>автор: <> </>
                         {props.mnemonic.creator.nickname}
                     </span>
+
+
                     <div className={s.buttonIcons}>
-                        <div onClick={pushBold}>  {boldIcon} </div>
-                        {!props.mnemonic.isCreator &&
+
+                        <div className={s.boldIconBox}>
+                            <div className={s.prompt}> показать мнемо-часть </div>
+                            <div onClick={pushBold}>{boldIcon}</div>
+                        </div>
+
+                        {!props.mnemonic.isCreator && props.auth &&
                         <div onClick={addMeMnemonic}> {addDelete} </div>
                         }
-                        {props.mnemonic.isCreator && !props.mnemonic.exampleExists &&
-                            <div onClick={deleteMnemonic} className={s2.deleteIcon}><Trash/></div>
+                        {props.mnemonic.isCreator && !props.mnemonic.exampleExists && props.auth &&
+                            <div onClick={showDeleteModal} className={s2.deleteIcon}><Trash/></div>
                         }
-                        {props.mnemonic.isCreator && !props.mnemonic.exampleExists &&
+                        {props.mnemonic.isCreator && !props.mnemonic.exampleExists && props.auth &&
                             <div onClick={editMnemo}> <Pencil/> </div>
                         }
                         <div onClick={addLikeMnemonic}>{Like}</div>
@@ -112,9 +155,19 @@ const MnemonicComponent: FC<PropsType> = (props) => {
                         mnemonicId={props.mnemonic.mnemonicId}
                         mnemonicPhrase={props.mnemonic.phrase}
                         setEdit={setEdit}
+                        engWord ={props.engWord}
                     />
                 </>
             }
+            <DeleteModal
+                classElement = {"мнемонику"}
+                elementMnemonic={props.mnemonic}
+                elementExample={null}
+                elementMyExample ={null}
+                show={showModal}
+                close={closeDeleteModal}
+                delete={deleteMnemonic}
+            />
         </div>
